Mushroom UC secured, sorting in mushroom list changed

+++ mushroomhunter/mushroomhunter-Web/src/main/java/cz/fi/muni/pa165/mushroomhunter/rest/MushroomRest.java
@@ -10,12 +10,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import cz.fi.muni.pa165.mushroomhunter.api.dto.MushroomDto;
 import cz.fi.muni.pa165.mushroomhunter.api.service.MushroomService;
+import cz.fi.muni.pa165.mushroomhunter.api.service.SecurityService;
 import java.util.ArrayList;
 import java.util.List;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.access.AccessDeniedException;
 
 /**
  *
@@ -27,6 +29,9 @@ public class MushroomRest {
 
     @Autowired
     MushroomService mushroomService;
+    
+    @Autowired
+    SecurityService securityService;
 
     @RequestMapping(method = RequestMethod.GET)
     public List<MushroomDto> getMushroomList() {
@@ -51,12 +56,20 @@ public class MushroomRest {
 
     @RequestMapping(method = RequestMethod.PUT)
     public MushroomDto updateMushroom(@RequestBody @Valid MushroomDto mushroom) {
+        Long currentUserId = securityService.getCurrentlyLoggedUser().getId();
+        if (!securityService.isCurrentlyLoggedUserAdmin()){
+            throw new AccessDeniedException("Access denien: User "+currentUserId+" is not admin so he cannot update mushroom "+mushroom.getId());
+        }
         return mushroomService.update(mushroom);
     }
 
     @RequestMapping(value = "{mushroomId}", method = RequestMethod.DELETE)
     public void deleteMushroom(@PathVariable Long mushroomId) {
         MushroomDto mushroom = mushroomService.find(mushroomId);
+        Long currentUserId = securityService.getCurrentlyLoggedUser().getId();
+        if (!securityService.isCurrentlyLoggedUserAdmin()){
+            throw new AccessDeniedException("Access denien: User "+currentUserId+" is not admin so he cannot delete mushroom "+mushroom.getId());
+        }
         mushroomService.delete(mushroom);
     }
